refactor: adjust imports, create separate dtos, remove unused classes
