+++ user-service/src/main/java/com/financialmonitoring/userservice/adapter/mapper/AuthMapper.java
@@ -1,0 +1,14 @@
+package com.financialmonitoring.userservice.adapter.mapper;
+
+import com.financialmonitoring.userservice.adapter.dto.UserLoginResponseDTO;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthMapper {
+
+
+    public UserLoginResponseDTO toUserLoginResponseDTO(UserLoginResponseDTO userLoginResponseDTO) {
+
+    }
+
+}
