+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/in/LoginUseCase.java
@@ -1,7 +1,7 @@
 package com.financialmonitoring.userservice.domain.port.in;
 
-import com.financialmonitoring.userservice.domain.dto.LoginRequestDTO;
-import com.financialmonitoring.userservice.domain.dto.LoginResponseDTO;
+import com.financialmonitoring.userservice.adapter.dto.LoginRequestDTO;
+import com.financialmonitoring.userservice.adapter.dto.LoginResponseDTO;
 
 public interface LoginUseCase {
     LoginResponseDTO login(LoginRequestDTO requestDTO);
