+++ user-service/src/main/java/com/financialmonitoring/userservice/domain/port/in/RegisterUserUseCase.java
@@ -1,7 +1,7 @@
 package com.financialmonitoring.userservice.domain.port.in;
 
-import com.financialmonitoring.userservice.domain.dto.RegisterRequestDTO;
-import com.financialmonitoring.userservice.domain.dto.RegisterResponseDTO;
+import com.financialmonitoring.userservice.adapter.dto.RegisterRequestDTO;
+import com.financialmonitoring.userservice.adapter.dto.RegisterResponseDTO;
 
 public interface RegisterUserUseCase {
     RegisterResponseDTO register(RegisterRequestDTO requestDTO);
